requirements.txt
# Web Framework
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.0.6
jinja2==3.1.2

# Database
sqlalchemy>=1.4.0,<1.5.0
alembic==1.12.1

# ML & Data Processing
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3

# Development
black==23.11.0
pylint==3.0.2
pytest==7.4.3
python-dotenv>=0.19.0,<0.20.0

# Utils
pydantic>=1.8.0,<2.0.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0 